Lab Project 3 (Data Clumps Example 1) - Description: The flaw is a data clump where a multiple variables are passed throughout the project, from class to class.
- Method:
For refactoring this flaw, I used an existing data class which is the Debts.kt data class and assign the data clump at the AddDebtDetails.kt class where it is started to be passed. I choose this method as it is most convenient to pass multiple data as a data class rather than typing it one by one.

- Conclusion:
Refactoring this flaw improve code readability, maintainability, and reduces duplication. It also makes it easier to modify or extend the data class in the future, as changes only need to be made in one place.