Lab Project 3 (Temporary Field Example 3) - Description: The code flaw is an unnecessary temporary field where the TextFields are declared into a variable beforehand.
- Method:
For refactoring this flaw, I deleted the temporary variables and made them inlined in the all fields checking function. I used this method as it is unnecessary to declare the temporary variables before.

- Conclusion:
Refactoring this flaw improve code clarity and its maintainability. Readability for future developments will be improved and easier to modify.