Lab Project 3 (Long Method Example 1) - Description: The code flaw is a long method in an adapter where the onBindViewHolder had long lines of UI updates.
- Method:
For refactoring this flaw, I use the extract method and extracted the lines of code with same conditions into an extracted function with more clear function names and added comments. I use this method to improve the readability of the code.

- Conclusion:
With lesser lines of code and clear function names, the code is more readable and easier to maintain or updated for future developments.