Lab Project 3 (Conditional Complexity Example 1) - Description: The code flaw is a method where multiple ifs and unclear conditions are shown in a RecyclerView Adapter. The code was also made more complex with a long method (which is already refactored before this commit)
- Method:
For refactoring this flaw, I delegated some of the nested if into a seperate function and added a new clearer variables and conditions to check the data being passed. I use this method to further improve the readability of the code.

- Conclusion:
With more comments and clearer conditions, the complexity of the code is highly reduced and is easier to read and modify for further development.