Lab Project 3 (Primitive Obsession Example 1) - Description: The code flaw was about a class dateInteger to show the debt from the most passed date to the furthest date in the future and it used an additional date in integer type.
- Method:
For refactoring this flaw, I used the date string variable to set the list into a sorted list where the most passed date at the top of the list and the furthest date in the future to be the most bottom using the funstions in the utils.kt class. I used this method so i can unuse the dateInteger data type

- Conclusion:
Processing the string type date will makes it less data type used and makes the data more understandable. (I have tried deleting the dateInteger column from the built table but it keeps on giving errors I don't understand so I just let the dateInteger into 0)